refactor: Update entity pages and playlists
