+++ frontend/js/src/user/components/PinnedRecordingCard.tsx
@@ -3,15 +3,15 @@ import { faThumbtack } from "@fortawesome/free-solid-svg-icons";
 import { IconProp } from "@fortawesome/fontawesome-svg-core";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { toast } from "react-toastify";
-import ListenCard from "../listens/ListenCard";
-import ListenControl from "../listens/ListenControl";
-import { ToastMsg } from "../notifications/Notifications";
-import GlobalAppContext from "../utils/GlobalAppContext";
+import ListenCard from "../../listens/ListenCard";
+import ListenControl from "../../listens/ListenControl";
+import { ToastMsg } from "../../notifications/Notifications";
+import GlobalAppContext from "../../utils/GlobalAppContext";
 import {
   getArtistName,
   getTrackName,
   pinnedRecordingToListen,
-} from "../utils/utils";
+} from "../../utils/utils";
 
 export type PinnedRecordingCardProps = {
   pinnedRecording: PinnedRecording;
